requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-auto-prefetching==0.2.11
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jwt==1.3.1
mysql==0.0.3
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
